extracts the routes to their own file
